+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 
 ## [Unreleased]
 
+### Added
+
+- Allow configuring the behavior for failures in method argument resolution (mostly deserialization) for
+  `@JetStreamListener`-annotated methods.
+
 ## [0.3.0] - 2026-06-13
 
 ### Breaking